# docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Finance API",
        "description": "Personal Finance Management API",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000/api",
            "description": "Local development server"
        }
    ],
    "paths": {
...
        "/api/exchange": {
            "get": {
                "tags": [
                    "Exchange"
                ],
                "summary": "List all supported currencies and rates",
                "operationId": "2765d24334f93d8eb96cbfcec1c99d41",
                "responses": {
                    "200": {
                        "description": "List of currencies and rates"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/exchange/convert": {
            "get": {
                "tags": [
                    "Exchange"
                ],
                "summary": "Convert an amount from one currency to another",
                "operationId": "06ff49f53212ac4b35612b992b9f4c7a",
                "parameters": [
                    {
                        "name": "amount",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "number"
                        }
                    },
                    {
                        "name": "from",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "to",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Conversion result"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/exchange/transactions": {
            "post": {
                "tags": [
                    "Exchange"
                ],
                "summary": "Convert multiple transactions to a target currency",
                "operationId": "4fbb32ce4e1dc27945645d0eb9987f6b",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "transactions",
                                    "target_currency"
                                ],
                                "properties": {
                                    "transactions": {
                                        "type": "array",
                                        "items": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer"
                                                },
                                                "amount": {
                                                    "type": "number"
                                                },
                                                "currency": {
                                                    "type": "string"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "target_currency": {
                                        "type": "string",
                                        "example": "GBP"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Transactions converted"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/notifications": {
            "get": {
                "tags": [
                    "Notifications"
                ],
                "summary": "List all notifications for the authenticated user",
                "operationId": "911565ad9aff9ba7eded3f189509e578",
                "responses": {
                    "200": {
                        "description": "Notifications list returned"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/notifications/unread": {
            "get": {
                "tags": [
                    "Notifications"
                ],
                "summary": "List only unread notifications",
                "operationId": "bbdf4779601a9c7758ebf68b0aedeb83",
                "responses": {
                    "200": {
                        "description": "Unread notifications list returned"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/notifications/{id}/read": {
            "patch": {
                "tags": [
                    "Notifications"
                ],
                "summary": "Mark a notification as read",
                "operationId": "65581bb40eba3764fb900f21495814de",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Notification marked as read"
                    },
                    "404": {
                        "description": "Notification not found"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/notifications/unread/count": {
            "get": {
                "tags": [
                    "Notifications"
                ],
                "summary": "Get the number of unread notifications",
                "operationId": "723bc14fc7dede3524d842ddd1198458",
                "responses": {
                    "200": {
                        "description": "Unread notifications count returned"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
        "/api/transactions": {
            "get": {
                "tags": [
                    "Transactions"
                ],
                "summary": "List transactions",
                "description": "Retrieve a paginated list of the authenticated user's transactions.",
                "operationId": "6581469356772e7a74a0d79897544f1a",
                "parameters": [
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Number of results per page",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 10
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of transactions",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "description": {
                                                        "type": "string",
                                                        "example": "Grocery shopping"
                                                    },
                                                    "amount": {
                                                        "type": "number",
                                                        "format": "float",
                                                        "example": 125.5
                                                    },
                                                    "currency": {
                                                        "type": "string",
                                                        "example": "GBP"
                                                    },
                                                    "type": {
                                                        "type": "string",
                                                        "enum": [
                                                            "income",
                                                            "expense"
                                                        ],
                                                        "example": "expense"
                                                    },
                                                    "date": {
                                                        "type": "string",
                                                        "format": "date",
                                                        "example": "2025-09-08"
                                                    },
                                                    "category_id": {
                                                        "type": "integer",
                                                        "example": 2
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Transaction list"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Access denied"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Transactions"
                ],
                "summary": "Create a new transaction",
                "operationId": "ab765dced58cfe2f620929db31452145",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "description",
                                    "amount",
                                    "type",
                                    "date",
                                    "category_id"
                                ],
                                "properties": {
                                    "description": {
                                        "type": "string",
                                        "example": "Grocery shopping"
                                    },
                                    "amount": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 120.5
                                    },
                                    "currency": {
                                        "type": "string",
                                        "example": "USD"
                                    },
                                    "type": {
                                        "type": "string",
                                        "enum": [
                                            "income",
                                            "expense"
                                        ],
                                        "example": "expense"
                                    },
                                    "date": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2025-09-08"
                                    },
                                    "category_id": {
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Transaction created successfully"
                    },
                    "400": {
                        "description": "Validation error"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
...
}